reorganize files into folder
