+++ client/src/App.js
@@ -1,9 +1,9 @@
 import "./App.css";
 import NavBar from "./components/nav-bar";
-import HomePage from "./components/HomePage";
-import SearchPage from "./components/SearchPage";
-import Recommendation from "./components/Recommendation";
-import ProfilePage from "./components/ProfilePage";
+import HomePage from "./components/HomePage/HomePage";
+import SearchPage from "./components/SearchPage/SearchPage";
+import Recommendation from "./components/RecomPage/Recommendation";
+import ProfilePage from "./components/ProfilePage/ProfilePage";
 //import Profile from "./components/profile";
 
 import { useAuth0 } from '@auth0/auth0-react';
