+++ client/src/components/SearchPage/SearchBar.js
